app.py
import streamlit as st
import pandas as pd
import plotly.express as px

# Assuming df_clubs is your DataFrame
# Load your data or replace this with your actual data loading code
df_clubs = pd.read_csv("./data/clubs.csv")

# Setting the app title and description
st.title("Club Analysis Dashboard")
st.sidebar.header("Settings")

# Sidebar: Club Selection
selected_clubs = st.sidebar.multiselect("Select Clubs", df_clubs['name'].unique())

# Sidebar: Metrics Selection
selected_metrics = st.sidebar.multiselect(
    "Select Metrics",
    ["total_market_value", "squad_size", "average_age", "foreigners_percentage"]
)

# Filter the DataFrame based on user selection
filtered_df = df_clubs[df_clubs['name'].isin(selected_clubs)]

# Overview Section
st.header("Overview")
st.write("Summary statistics for selected clubs:")
st.write(filtered_df[selected_metrics].describe())

# Comparison Section
st.header("Club Comparison")
if selected_clubs:
    # Plotly Express Scatter Plot for comparison
    fig = px.scatter(filtered_df, x="squad_size", y="total_market_value", color="name", size="average_age", title="Club Comparison")
    st.plotly_chart(fig)
else:
    st.warning("Select at least one club for comparison.")

# Seasonal Performance Section
st.header("Seasonal Performance")
# Assuming you have a DataFrame df_seasonal_performance with seasonal data
# Display a line chart or other relevant visualization

# Geographical Data Section
st.header("Geographical Data")
# Assuming you have a DataFrame df_geographical_data with geographical data
# Display a map or other relevant visualization

# Transfer Market Analysis Section
st.header("Transfer Market Analysis")
# Assuming you have a DataFrame df_transfer_market_analysis with transfer data
# Display a bar chart or other relevant visualization

# Coach Analysis Section
st.header("Coach Analysis")
# Assuming you have a DataFrame df_coach_analysis with coach data
# Display relevant coach information

# Add more sections as needed...

# Finalize the app
