requirements.txt
numpy
pandas
streamlit
scikit-learn
matplotlib
plotly.express